Refactor naming and restructure
